m2-pdf-processing/requirements.txt
Flask==2.3.2
gunicorn==20.1.0
numpy==2.1.1
Pillow==10.4.0
matplotlib==3.9.2
pdf2image==1.17.0
requests==2.32.3
